+++ react/components/customers-admin.tsx
@@ -4,11 +4,11 @@ import React, { useState } from 'react'
 import type { WrappedComponentProps } from 'react-intl'
 import { injectIntl, defineMessages } from 'react-intl'
 import { useQuery, useMutation } from 'react-apollo'
-import { Button, Dropdown, Toggle } from 'vtex.styleguide'
+import { Button, Dropdown, Toggle, Alert } from 'vtex.styleguide'
 
-import GET_USER from '../../queries/getUser.gql'
-import GET_ROLES from '../../queries/ListRoles.gql'
-import SAVE_USER from '../../mutations/saveUser.gql'
+import GET_USER from '../queries/getUser.gql'
+import GET_ROLES from '../queries/ListRoles.gql'
+import SAVE_USER from '../mutations/saveUser.gql'
 
 const messages = defineMessages({
   role: {
@@ -39,26 +39,34 @@ const messages = defineMessages({
     id: 'admin/storefront-permissions.button.save',
     defaultMessage: 'Save',
   },
+  success: {
+    id: 'admin/storefront-permissions.tab.users.success',
+    defaultMessage: 'B2B info saved',
+  },
+  error: {
+    id: 'admin/storefront-permissions.tab.users.error',
+    defaultMessage: 'Error saving B2B info',
+  },
 })
 
 const UserEdit: FC<any & WrappedComponentProps> = (props: any) => {
-  const { intl } = props
-
-  const [saveUser, { loading: saveUserLoading }] = useMutation(SAVE_USER)
+  const { intl, id, showName, showEmail, showCancel, onCancel, onSave } = props
 
   const [state, setState] = useState<any>({
+    message: null,
     id: null,
     roleId: null,
     userId: null,
-    name: null,
-    email: null,
+    clId: id,
+    name: props.name ?? null,
+    email: props.email ?? null,
     canImpersonate: false,
   })
 
-  const { loading } = useQuery(GET_USER, {
-    skip: !props?.id,
+  const { loading, data: getUserData } = useQuery(GET_USER, {
+    skip: !id,
     variables: {
-      id: props?.id,
+      id,
     },
     fetchPolicy: 'network-only',
     onCompleted: (res: any) => {
@@ -69,29 +77,53 @@ const UserEdit: FC<any & WrappedComponentProps> = (props: any) => {
     },
   })
 
+  const [saveUser, { loading: saveUserLoading }] = useMutation(SAVE_USER, {
+    onCompleted(res: any) {
+      if (onSave) {
+        onSave()
+      }
+
+      setState({
+        ...state,
+        id: state.id ?? res?.saveUser?.id,
+        message: 'success',
+      })
+    },
+
+    onError: () => {
+      setState({
+        ...state,
+        message: 'error',
+      })
+    },
+  })
+
   const { loading: loadingRoles, data: dataRoles } = useQuery(GET_ROLES)
 
   const handleSaveUser = () => {
+    const variables: any = {
+      id: state.id,
+      clId: state.clId,
+      userId: state.userId,
+      roleId: state.roleId,
+      name: state.name,
+      email: state.email,
+      canImpersonate: state.canImpersonate,
+    }
+
+    if (state.id) {
+      variables.id = state.id
+    }
+
     saveUser({
-      variables: {
-        id: state.id,
-        userId: state.userId,
-        roleId: state.roleId,
-        name: state.name,
-        email: state.email,
-        canImpersonate: state.canImpersonate,
-      },
+      variables,
     })
   }
 
-  if (!props.id) return null
-
   return (
     <div className="w-100 pt6">
-      <div className="mb5">{state.name}</div>
-
-      <div className="mb5">{state.email}</div>
-
+      {showName && <div className="mb5">{state.name}</div>}
+      {showEmail && <div className="mb5">{state.email}</div>}
       <div className="mb5">
         <Dropdown
           label={intl.formatMessage(messages.role)}
@@ -123,9 +155,27 @@ const UserEdit: FC<any & WrappedComponentProps> = (props: any) => {
       </div>
 
       <div className="mv4 flex justify-between">
+        {showCancel && onCancel && (
+          <Button
+            variation="tertiary"
+            disabled={loading}
+            collapseLeft
+            onClick={() => {
+              onCancel()
+            }}
+          >
+            {intl.formatMessage(messages.cancel)}
+          </Button>
+        )}
         <Button
           variation="primary"
-          disabled={loading || saveUserLoading || !state.name || !state.email}
+          disabled={
+            loading ||
+            saveUserLoading ||
+            !state.name ||
+            !state.email ||
+            !state.roleId
+          }
           collapseRight
           onClick={() => {
             handleSaveUser()
@@ -134,6 +184,18 @@ const UserEdit: FC<any & WrappedComponentProps> = (props: any) => {
           {intl.formatMessage(messages.save)}
         </Button>
       </div>
+      {state.message && (
+        <Alert
+          type={state.message}
+          onClose={() => {
+            setState({ ...state, message: null })
+          }}
+        >
+          {state.message === 'success'
+            ? intl.formatMessage(messages.success)
+            : intl.formatMessage(messages.error)}
+        </Alert>
+      )}
     </div>
   )
 }